requirements.txt
supabase-py
